re-struct project and add home screen skeleton
